+++ src/Kephas.Reflection/Reflection/DefaultTypeResolver.cs
@@ -16,6 +16,7 @@ namespace Kephas.Reflection
     using System.Linq;
     using System.Reflection;
 
+    using Kephas.Application;
     using Kephas.Injection.AttributedModel;
     using Kephas.Logging;
     using Kephas.Resources;
@@ -34,17 +35,20 @@ namespace Kephas.Reflection
         /// <summary>
         /// Initializes a new instance of the <see cref="DefaultTypeResolver"/> class.
         /// </summary>
-        /// <param name="ambientServices">The ambient services.</param>
+        /// <param name="appRuntime">The application runtime.</param>
         /// <param name="typeLoader">Optional. The type loader.</param>
         /// <param name="logManager">Optional. The log manager.</param>
         [InjectConstructor]
-        public DefaultTypeResolver(IAmbientServices ambientServices, ITypeLoader? typeLoader = null, ILogManager? logManager = null)
+        public DefaultTypeResolver(
+            IAppRuntime appRuntime,
+            ITypeLoader? typeLoader = null,
+            ILogManager? logManager = null)
             : base(logManager)
         {
-            ambientServices = ambientServices ?? throw new ArgumentNullException(nameof(ambientServices));
+            appRuntime = appRuntime ?? throw new ArgumentNullException(nameof(appRuntime));
 
             this.typeLoader = typeLoader ?? new DefaultTypeLoader(logManager);
-            this.getAppAssemblies = () => ambientServices.GetAppAssemblies();
+            this.getAppAssemblies = () => appRuntime.GetAppAssemblies();
         }
 
         /// <summary>