+++ src/Kephas.Workflow.Model/Runtime/ModelRegistries/WorkflowModelRegistry.cs
@@ -25,17 +25,17 @@ namespace Kephas.Workflow.Model.Runtime.ModelRegistries
         /// Initializes a new instance of the <see cref="WorkflowModelRegistry"/> class.
         /// </summary>
         /// <param name="injectableFactory">The injectable factory.</param>
-        /// <param name="ambientServices">The ambient services.</param>
+        /// <param name="appRuntime">The application runtime.</param>
         /// <param name="typeLoader">The type loader.</param>
         /// <param name="logManager">Optional. The log manager.</param>
         public WorkflowModelRegistry(
             IInjectableFactory injectableFactory,
-            IAmbientServices ambientServices,
+            IAppRuntime appRuntime,
             ITypeLoader typeLoader,
             ILogManager? logManager = null)
             : base(
                 injectableFactory,
-                ambientServices,
+                appRuntime,
                 typeLoader,
                 context =>
                 {