+++ src/Tests/Kephas.Model.Tests/Security/Permissions/Runtime/ModelRegistries/PermissionsModelRegistryTest.cs
@@ -34,7 +34,7 @@ namespace Kephas.Model.Tests.Security.Permissions.Runtime.ModelRegistries
         [Test]
         public async Task GetRuntimeElementsAsync()
         {
-            var appRuntime = Substitute.For<IAmbientServices>();
+            var appRuntime = Substitute.For<IAppRuntime>();
             appRuntime
                 .GetAppAssemblies()
                 .Returns(new[] { this.GetType().Assembly });
@@ -69,7 +69,7 @@ namespace Kephas.Model.Tests.Security.Permissions.Runtime.ModelRegistries
         [Test]
         public async Task GetRuntimeElementsAsync_ExcludeFromModel()
         {
-            var appRuntime = Substitute.For<IAmbientServices>();
+            var appRuntime = Substitute.For<IAppRuntime>();
             appRuntime
                 .GetAppAssemblies()
                 .Returns(new[] { this.GetType().Assembly });